fix(standard-cli): address 6 unresolved PR #921 review comments
  - Fix --network flag being overwritten by initApiCli() reading
  config.conf;
    user's explicit choice now takes precedence (Contract 1 violation)
  - Fix --grpc-endpoint not syncing network metadata; now auto-detects
    known endpoints, validates against --network, falls back to CUSTOM
  - Fix intermediate char[] from toCharArray() not cleared in
  authenticate()
  - Add ParsedOptions.getInt() with overflow check; replace all 19
  unsafe
    (int) opts.getLong() casts across 6 command files (Contract 2
  violation)
  - Add CommandSupport.requirePositive() and apply amount > 0
  validation
    to 10 transaction/staking handlers; add asset-issue pre-validation
   to
    prevent legacy System.out.println pollution in JSON mode
  (Contracts 2,4,5)
  - Wrap Long.parseLong() in create-proposal with try-catch and
  usageError
    (Contract 2 violation)

+++ src/main/java/org/tron/walletcli/cli/commands/ContractCommands.java
@@ -122,7 +122,7 @@ public class ContractCommands {
                     long callValue = opts.has("value") ? opts.getLong("value") : 0;
                     long tokenValue = opts.has("token-value") ? opts.getLong("token-value") : 0;
                     String tokenId = opts.has("token-id") ? opts.getString("token-id") : "";
-                    int permissionId = opts.has("permission-id") ? (int) opts.getLong("permission-id") : 0;
+                    int permissionId = opts.has("permission-id") ? opts.getInt("permission-id") : 0;
                     boolean multi = opts.getBoolean("multi");
 
                     byte[] data = ByteArray.fromHexString(AbiUtil.parseMethod(method, params, false));
